extract PageviewSlider component
